test(review): split marker packaging coverage
