+++ tests/validation/test_validate_review_marker.py
@@ -23,14 +23,6 @@ import pytest
 
 REPO_ROOT = Path(__file__).resolve().parent.parent.parent
 SCRIPT_PATH = REPO_ROOT / "scripts" / "validation" / "validate_review_marker.py"
-SKILL_SCRIPT_PATH = (
-    REPO_ROOT / ".claude" / "skills" / "review" / "scripts" / "validate_review_marker.py"
-)
-COPILOT_SCRIPT_PATH = (
-    REPO_ROOT / "src" / "copilot-cli" / "skills" / "review" / "scripts" / "validate_review_marker.py"
-)
-SHIP_COMMAND_PATH = REPO_ROOT / ".claude" / "commands" / "ship.md"
-COPILOT_SHIP_SKILL_PATH = REPO_ROOT / "src" / "copilot-cli" / "skills" / "ship" / "SKILL.md"
 
 
 def _load_module():
@@ -384,38 +376,6 @@ def test_source_validator_cli_runs() -> None:
     )
     assert result.returncode == 0
     assert "Validate a SHA-bound" in result.stdout
-
-
-def test_copilot_plugin_contains_review_validator() -> None:
-    """Vendored plugin installs include the same validator under the review skill."""
-    source_text = SCRIPT_PATH.read_text(encoding="utf-8")
-    assert SKILL_SCRIPT_PATH.read_text(encoding="utf-8") == source_text
-    assert COPILOT_SCRIPT_PATH.read_text(encoding="utf-8") == SKILL_SCRIPT_PATH.read_text(
-        encoding="utf-8"
-    )
-
-
-def test_ship_docs_use_review_skill_validator_paths() -> None:
-    """Source and Copilot /ship docs point at plugin-shipped validator paths."""
-    for path in (SHIP_COMMAND_PATH, COPILOT_SHIP_SKILL_PATH):
-        text = path.read_text(encoding="utf-8")
-        assert "review/scripts/validate_review_marker.py" in text
-        assert '--repo-root "$(pwd)"' in text
-        assert "scripts/validation/validate_review_marker.py --ref HEAD" not in text
-
-
-def test_review_marker_docs_use_repo_resolvable_references() -> None:
-    """Shipped docs and scripts do not point at private memory identifiers."""
-    private_ref = "decision-review-marker-sha-binding-mechanism"
-    paths = [
-        SCRIPT_PATH,
-        SKILL_SCRIPT_PATH,
-        COPILOT_SCRIPT_PATH,
-        REPO_ROOT / ".claude" / "skills" / "review" / "SKILL.md",
-        REPO_ROOT / "src" / "copilot-cli" / "skills" / "review" / "SKILL.md",
-    ]
-    for path in paths:
-        assert private_ref not in path.read_text(encoding="utf-8")
 
 
 # --- validate_ref: edge -----------------------------------------------------
